* generic/tclStringObj.c:       Changed type of the 'allocated' field         * generic/tclTestObj.c:         of the String struct (and the         TestString counterpart) from size_t to int since only int values are         ever stored in it.

+++ generic/tclTestObj.c
@@ -55,7 +55,7 @@ static int		TeststringobjCmd(ClientData dummy, Tcl_Interp *interp,
 
 typedef struct TestString {
     int numChars;
-    size_t allocated;
+    int allocated;
     size_t uallocated;
     Tcl_UniChar unicode[2];
 } TestString;
